+++ crates/web-client/js/node-index.js
@@ -157,6 +157,7 @@ export const AccountDelta = /* @__PURE__ */ _reexport("AccountDelta");
 export const AccountFile = /* @__PURE__ */ _reexport("AccountFile");
 export const AccountHeader = /* @__PURE__ */ _reexport("AccountHeader");
 export const AccountId = /* @__PURE__ */ _reexport("AccountId");
+export const AccountInputs = /* @__PURE__ */ _reexport("AccountInputs");
 export const AccountInterface = /* @__PURE__ */ _reexport("AccountInterface");
 export const AccountPatch = /* @__PURE__ */ _reexport("AccountPatch");
 export const AccountProof = /* @__PURE__ */ _reexport("AccountProof");
@@ -215,6 +216,7 @@ export const GetProceduresResultItem = /* @__PURE__ */ _reexport(
   "GetProceduresResultItem"
 );
 export const InputNote = /* @__PURE__ */ _reexport("InputNote");
+export const InputNoteAndArgs = /* @__PURE__ */ _reexport("InputNoteAndArgs");
 export const InputNoteRecord = /* @__PURE__ */ _reexport("InputNoteRecord");
 export const InputNoteState = /* @__PURE__ */ _reexport("InputNoteState");
 export const InputNotes = /* @__PURE__ */ _reexport("InputNotes");