+++ crates/web-client/src/models/mod.rs
@@ -36,6 +36,7 @@ pub mod account_delta;
 pub mod account_file;
 pub mod account_header;
 pub mod account_id;
+pub mod account_inputs;
 pub mod account_patch;
 pub mod account_proof;
 pub mod account_reader;
@@ -140,6 +141,8 @@ declare_js_miden_arrays! {
     (crate::models::storage_slot::StorageSlot) -> StorageSlotArray,
     (crate::models::transaction_script_inputs::TransactionScriptInputPair) -> TransactionScriptInputPairArray,
     (crate::models::felt::Felt) -> FeltArray,
+    (crate::models::account_inputs::AccountInputs) -> AccountInputsArray,
+    (crate::models::transaction_request::input_note_and_args::InputNoteAndArgs) -> InputNoteAndArgsArray,
     (crate::models::transaction_request::note_and_args::NoteAndArgs) -> NoteAndArgsArray,
     (crate::models::transaction_request::note_details_and_tag::NoteDetailsAndTag) -> NoteDetailsAndTagArray,
     (crate::models::transaction_request::note_id_and_args::NoteIdAndArgs) -> NoteIdAndArgsArray